# docs/openapi.yaml
openapi: "3.0.3"

info:
  title: NAEOS API
  description: Neural Architecture Engine for Orchestration & Synthesis
  version: "0.5.0"

servers:
  - url: /api/v1

paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /specs:
    get:
      operationId: listSpecs
      summary: List all specifications
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of specs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecListResponse"
    post:
      operationId: createSpec
      summary: Create a new specification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecCreateRequest"
      responses:
        "201":
          description: Spec created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /specs/validate:
    post:
      operationId: validateSpec
      summary: Validate a specification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecValidateRequest"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecValidateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /specs/compile:
    post:
      operationId: compileSpec
      summary: Compile a specification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecCompileRequest"
      responses:
        "200":
          description: Compilation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpecCompileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /pipeline/run:
    post:
      operationId: runPipeline
      summary: Execute a pipeline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineRunRequest"
      responses:
        "202":
          description: Pipeline accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRunResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /pipeline/status:
    get:
      operationId: getPipelineStatus
      summary: Get pipeline execution status
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pipeline status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  /artifacts:
    get:
      operationId: listArtifacts
      summary: List artifacts
      parameters:
        - name: spec_id
          in: query
          schema:
            type: string
        - name: kind
          in: query
          schema:
            type: string
            enum: [source, binary, manifest, log]
      responses:
        "200":
          description: List of artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactListResponse"
    post:
      operationId: uploadArtifact
      summary: Upload an artifact
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, kind]
              properties:
                file:
                  type: string
                  format: binary
                kind:
                  type: string
                  enum: [source, binary, manifest, log]
                spec_id:
                  type: string
      responses:
        "201":
          description: Artifact uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /context/generate:
    post:
      operationId: generateContext
      summary: Generate context for a spec
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContextGenerateRequest"
      responses:
        "200":
          description: Generated context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContextGenerateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /mcp/message:
    post:
      operationId: mcpMessage
      summary: Send an MCP message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MCPMessageRequest"
      responses:
        "200":
          description: MCP response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MCPMessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        version:
          type: string
          example: "0.5.0"

    SpecCreateRequest:
      type: object
      required: [name, content]
      properties:
        name:
          type: string
        content:
          type: object
        version:
          type: string
          default: "1.0.0"
        description:
          type: string

    SpecResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        content:
          type: object
        created_at:
          type: string
          format: date-time

    SpecListResponse:
      type: object
      properties:
        specs:
          type: array
          items:
            $ref: "#/components/schemas/SpecResponse"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    SpecValidateRequest:
      type: object
      required: [spec_id]
      properties:
        spec_id:
          type: string

    SpecValidateResponse:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"

    ValidationError:
      type: object
      properties:
        path:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [error, warning, info]

    SpecCompileRequest:
      type: object
      required: [spec_id]
      properties:
        spec_id:
          type: string
        target:
          type: string
          enum: [go, wasm, tflite]
          default: go

    SpecCompileResponse:
      type: object
      properties:
        artifact_id:
          type: string
        status:
          type: string
          example: compiled

    PipelineRunRequest:
      type: object
      required: [spec_id, pipeline]
      properties:
        spec_id:
          type: string
        pipeline:
          type: string
          enum: [build, deploy, test, full]
        params:
          type: object
          additionalProperties: true

    PipelineRunResponse:
      type: object
      properties:
        pipeline_id:
          type: string
        status:
          type: string
          example: queued

    PipelineStatusResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/PipelineStep"
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time

    PipelineStep:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [pending, running, succeeded, failed, skipped]
        message:
          type: string

    ArtifactResponse:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        spec_id:
          type: string
        size:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time

    ArtifactListResponse:
      type: object
      properties:
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactResponse"
        total:
          type: integer

    ContextGenerateRequest:
      type: object
      required: [spec_id]
      properties:
        spec_id:
          type: string
        context_type:
          type: string
          enum: [full, summary, dependencies]
          default: full

    ContextGenerateResponse:
      type: object
      properties:
        context_id:
          type: string
        context:
          type: object
        generated_at:
          type: string
          format: date-time

    MCPMessageRequest:
      type: object
      required: [method, params]
      properties:
        jsonrpc:
          type: string
          default: "2.0"
        method:
          type: string
        params:
          type: object
        id:
          oneOf:
            - type: string
            - type: integer

    MCPMessageResponse:
      type: object
      properties:
        jsonrpc:
          type: string
          example: "2.0"
        result:
          type: object
        error:
          type: object
          nullable: true
          properties:
            code:
              type: integer
            message:
              type: string
        id:
          oneOf:
            - type: string
            - type: integer

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object